requirements.txt
amqp
asgiref==3.7.2
atomicwrites==1.4.1
apscheduler
aiohttp
celery==5.3.6

dj-database-url==2.0.0
dj-rest-auth==2.2.7
dj-static==0.0.6
Django==4.2.4

django-cors-headers==4.2.0
django-redis
django-email-verification==0.3.1
django-filter==23.1
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.1.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-crontab
drf-writable-nested
drf-yasg==1.21.7
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
flower
gcloud==0.18.3
gunicorn==21.2.0

MT5Manager

packaging==23.1
phonenumberslite==8.13.20
Pillow==10.0.0
# psycopg2==2.9.6
# psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
pytest==7.1.1
python-multipart
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==4.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
pyotp
pamqp
pika
qrcode
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
setuptools
stripe
psycopg2-binary

static3==0.7.0
swagger-ui-bundle==0.0.9
urllib3==1.26.16
vine==5.1.0
waitress
wcwidth==0.2.12
whitenoise==6.5.0
metaapi-cloud-sdk
uvicorn
fastapi
channels
channels_redis
daphne

concurrent-log-handler
confluent-kafka
